Add "indexed" visitors
This patch adds a new set of visitor functions that include the index of
the field within the struct, in addition to its name.  The indices are
0-based, and the field ordering is the same as the ordering of the
parameters passed to the VISITABLE_STRUCT macro.

We also provide two new helpers in the visit_struct namespace.  One
returns the number of fields in the struct, and the other gives you
access to the field index enum.