rethink msvc extended constexpr options
This reverts to the state after 2d30107de2ec7e4220db1f9938f31480df71893c

At that state, CXX14 constexpr was enabled for MSVC 2017, and the
basic tests were compiling -- the visit_struct header was fine.

The CXX14 tests, which use visitors with mutable state in constant
expressions, are not passing. But this doesn't mean we should prevent
anyone on MSVC 2017 from getting a constexpr declaration on apply_visitor.
Maybe they can still make use of it in some useful way.